refactor: dissolve native_handler.dart tree and co-locate stubs at the leaf level
