+++ packages/logd/lib/logd.dart
@@ -2,11 +2,12 @@ library;
 
 export 'src/core/log_level.dart';
 export 'src/core/theme/log_theme.dart';
+export 'src/handler/engine/arena.dart';
+export 'src/handler/engine/native_engine.dart';
 export 'src/handler/handler.dart'
     hide DecoratorPipeline, PrintSink, RenderTokens, TerminalLayout;
-export 'src/handler/native_handler_stub.dart'
-    if (dart.library.io) 'src/handler/native_handler.dart'
-    hide BinaryAnsiEncoder, BinaryIR, BinaryIRWriter, BinaryToonEncoder;
+export 'src/handler/sink/file_sink.dart';
+export 'src/handler/sink/isolate_sink.dart';
 export 'src/logger/logger.dart' hide InternalLogger, LoggerCache;
 export 'src/stack_trace/stack_trace.dart' hide CallbackInfo, StackFrameSet;
 export 'src/time/timestamp.dart'
