+++ packages/logd/lib/src/handler/document/binary_ir_stub.dart
@@ -1,0 +1,14 @@
+library;
+
+class BinaryIR {
+  static const int alignLeft = 0;
+  static const int alignCenter = 1;
+  static const int alignRight = 2;
+  static const int alignJustify = 3;
+}
+
+class BinaryIRWriter {}
+
+class BinaryAnsiEncoder {}
+
+class BinaryToonEncoder {}
