--- packages/logd/lib/src/handler/native_handler.dart
@@ -1,35 +1,0 @@
-library;
-
-import 'dart:async';
-import 'dart:convert' as convert;
-import 'dart:ffi' as ffi;
-import 'dart:io' as io;
-import 'dart:isolate';
-import 'dart:math';
-import 'dart:typed_data';
-
-import 'package:characters/characters.dart';
-import 'package:ffi/ffi.dart' as pkg_ffi;
-import 'package:meta/meta.dart';
-
-import '../core/context/context.dart';
-import '../core/context/io/file_system.dart';
-import '../core/theme/log_theme.dart';
-import '../core/utils/utils.dart';
-import '../logger/logger.dart';
-import '../stack_trace/stack_trace.dart';
-import '../time/timestamp.dart';
-import 'handler.dart';
-
-part 'document/binary_ir.dart';
-part 'document/binary_ir_writer.dart';
-part 'document/native_packet.dart';
-part 'encoder/binary_ansi_encoder.dart';
-part 'encoder/binary_toon_encoder.dart';
-part 'engine/arena.dart';
-part 'engine/arena_engine.dart';
-part 'engine/native_engine.dart';
-part 'engine/native_isolate_worker.dart';
-part 'sink/file_sink.dart';
-part 'sink/isolate_sink.dart';
-part 'sink/native_isolate_sink.dart';
