+++ packages/logd/test/handler/binary_ir_test.dart
@@ -1,7 +1,9 @@
 import 'dart:ffi' as ffi;
 
-import 'package:logd/logd.dart' hide Arena;
-import 'package:logd/src/handler/native_handler.dart';
+import 'package:logd/logd.dart'
+    hide Arena, ArenaDocument, ArenaEngine, FileSink, IsolateSink, NativeEngine;
+import 'package:logd/src/handler/document/binary_ir_native.dart';
+import 'package:logd/src/handler/engine/arena_native.dart';
 import 'package:test/test.dart';
 
 void main() {
