+++ packages/logd/test/handler/document/binary_ir_v2_test.dart
@@ -1,7 +1,8 @@
 import 'dart:ffi' as ffi;
 
+import 'package:logd/src/handler/document/binary_ir_native.dart';
+import 'package:logd/src/handler/engine/arena_native.dart';
 import 'package:logd/src/handler/handler.dart';
-import 'package:logd/src/handler/native_handler.dart';
 import 'package:test/test.dart';
 
 void main() {
